move send notifications from init to its own module
